sample_projects/financial-transactions/pyproject.toml
[project]
name = "financial-transactions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "fastapi>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "redis>=6.0.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
